Add files via upload
FORTRAN Binary search
This is a re-write of an unattributed code Fortran IV segment found on Rosetta Code
https://rosettacode.org/wiki/Binary_search#Fortran. If someone can find the original author, let me know.
I cleaned the code up and restructured it to Fortran 90 (with F77 style logical operators)
It works fast and reliably but could probably use a further clean up. Note the use of the 
Logical flag to stop the loop instead of a simple Do and later EXIT statements. This is
because GFortran optimizes loops with an EXIT statement in them very poorly or not at all.
This construct allows the loop to be unrolled by the compiler.